[x-lint] move errors into their own module
