+++ devtools/x-lint/src/lib.rs
@@ -7,14 +7,16 @@
 //! [Arcanist](https://secure.phabricator.com/book/phabricator/article/arcanist_lint)'s lint engine.
 
 pub mod content;
+mod errors;
 pub mod file;
 pub mod package;
 pub mod project;
 mod runner;
 
+pub use errors::*;
 pub use runner::*;
 
-use std::{borrow::Cow, error, fmt, io, path::Path, process::ExitStatus, result};
+use std::{borrow::Cow, fmt, path::Path};
 
 /// Represents a linter.
 pub trait Linter: Send + Sync + fmt::Debug {
@@ -174,56 +176,6 @@ impl<'l> fmt::Display for LintKind<'l> {
             LintKind::File(path) => write!(f, "file {}", path.display()),
             LintKind::Content(path) => write!(f, "content {}", path.display()),
         }
-    }
-}
-
-/// Type alias for the return type for `run` methods.
-pub type Result<T, E = SystemError> = result::Result<T, E>;
-
-/// A system error that happened while running a lint.
-#[derive(Debug)]
-#[non_exhaustive]
-pub enum SystemError {
-    Exec {
-        cmd: &'static str,
-        status: ExitStatus,
-    },
-    Guppy(guppy::Error),
-    Io(io::Error),
-}
-
-impl fmt::Display for SystemError {
-    fn fmt(&self, f: &mut fmt::Formatter) -> fmt::Result {
-        match self {
-            SystemError::Exec { cmd, status } => match status.code() {
-                Some(code) => write!(f, "'{}' failed with exit code {}", cmd, code),
-                None => write!(f, "'{}' terminated by signal", cmd),
-            },
-            SystemError::Io(err) => write!(f, "IO error: {}", err),
-            SystemError::Guppy(err) => write!(f, "guppy error: {}", err),
-        }
-    }
-}
-
-impl error::Error for SystemError {
-    fn source(&self) -> Option<&(dyn error::Error + 'static)> {
-        match self {
-            SystemError::Exec { .. } => None,
-            SystemError::Io(err) => Some(err),
-            SystemError::Guppy(err) => Some(err),
-        }
-    }
-}
-
-impl From<guppy::Error> for SystemError {
-    fn from(err: guppy::Error) -> Self {
-        SystemError::Guppy(err)
-    }
-}
-
-impl From<io::Error> for SystemError {
-    fn from(err: io::Error) -> Self {
-        SystemError::Io(err)
     }
 }
